feat: 만료된 access token에 대한 재발급 처리 구현

+++ SlimMy/Model/MessageType.cs
@@ -6,7 +6,7 @@ using System.Threading.Tasks;
 
 namespace SlimMy.Model
 {
-    public enum MessageType : byte
+    public enum MessageType : ushort
     {
         UserJoinChatRoom = 1,
         ChatContent = 2,
